backend/pyproject.toml
[project]
name = "tasks-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "AleksandrZaec",email = "zaecam95@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "alembic (>=1.16.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (<4.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "sqladmin[async] (>=0.20.1,<0.21.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=6.2.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate -m 'auto'"
